feat: correct compute_gradient_mass_matrix so it makes sense
Previsouly `compute_gradient_mass_matrix` returned the sum of all contributions of gradient along some input dimension. This is not correct, as these contribute to different components of the gradient. Now the function instead returns only the contribution to the specified component.

There was also some simplification of the C code.