fix: fixted the covector Hodge
Covector basis Hodge previously did not take the sign into the account the sign of the basis.

I also added "hash" and "repr" magic methods, along with changing the "tp_richcompare" to only allow difference of sign for cases where operation is EQ or NE. Lastly, I added the method "normalize", which takes the sign off a basis. It is equivalent to calling "(+1, basis) if basis.sign > 1 else (-1, -basis)".